Clean up assert code
Bionic libc and GNU libc use different __assert() argument orders.
Previously, this was worked around by just macro defining __assert() on
host to be undefined. This causes a compilation error deep in glibc when
you try and actually compile with -DNDEBUG, as the macro definition ends
up clobbering the function declaration.

This code removes that build breakage, and patches up the asserts to
work both with glibc and bionic.

Bug: 160324350
Test: Build libbinder_ndk with -DNDEBUG && atest libbinder_ndk_unit_test binderVendorDoubleLoadTest
Change-Id: If4410e093abdce6197dcc960d79796c1ab17d7e9