Add a standardized shared memory type
This adds a simple SharedMemory type that can be used to pass blocks
of shared memory between processes over stable AIDL.

In addition a small utility library is provided for interfacing with
existing code that uses IMemory.

The intention is to add utilities for direct usage of the memory
content.

Bug: 139324991
Change-Id: I42dbd66b146a2257190ed766a5dac9361db93fa2