libbinder: APEX stability levels
FIXME: need a stability level between VNDK and APEX, called 'APEX_API'
which works between system/apex, but can't be used by vendor. This will
allow services to register with this API, avoid needing to be in the
VNDK manifest, but work system<->APEX.

Make sure that an unstable interface in an APEX can't leak to system or
vice versa.

Fixes: 139325195
Test: wip
Change-Id: Ia26a0baa7cee4a238b6fbdb0c049d2074d263638