Don't give SYSTEM stability to vendor APEXes
Previously, all APEXes including those are in the vendor partition got
the SYSTEM local stability. This will incorrectly allow non-stable
binders in the vendor APEXes to talk to non-stable binders in the system
partition.

To fix this, the SYSTEM local stability is allowed to 1) APEXes that are
built without VNDK (i.e. !defined(__ANDROID_AVNDK__)) and 2) the media
swcodec APEX which is the only system APEX that is built with VNDK.
All other APEXes built with VNDK gets the VENDOR local stability.

Note that this isn't the long-term solution. We will have per APEX
local stabilities to prevent a non-stable binder in an APEX from talking
to a non-stable binder in another APEX.

Bug: 142684679
Test: m
Change-Id: Ifd7c8bdf55848e4e890f87e3508dc3bf5b4fe881