kLocalStability is initialized in the source file
The static variable is used only within the libbinder. It's never
referenced outside of the library. However, it has been initialized in
the header conditionally with __ANDROID_VNDK__ and __ANDROID_APEX__.
This has cauesed some level of confusion that the variable is
initialized depending on the context wheure libbinder is used, which
actually isn't.

Bug: 161926892
Test: m libbinder
Test: watch TH for the binder tests
Change-Id: Ie8109c0e73b7f999de330b20cf9fb981c05dee49