SurfaceFlinger: implement crude headless support
Headless android will not boot fully because a working display is an
integral part of SurfaceFlinger. Make as minimal of changes as we can
get away with to work around this.

On the kernel-side, we're using the DRM_VKMS which, as of 4.19 has
very minimal virtual display support. VKMS is missing rendering
features that Android relies on, and seems to have incomplete
VSYNC generation support. Fortunately, everything appears to work
just fine if we simply tell SurfaceFlinger to keep this display
turned off and force use of SurfaceFlinger's synthetic vsync. VKMS
is still used, but with the display turned off, we avoid being
impacted by its missing features.

Headless mode is made available via an Android property:
`ro.surface_flinger.use_headless_mode`

This change was inspired but the approach taken here:
https://github.com/opersys/bbb-platform_frameworks_native/tree/bbb-hack

Test: booted bluejay-userdebug with the property set
Change-Id: I0f099fa99c650dfda14b249faca6050518e5b1ae