Suppress warning on the use of deprecated APIs
AChoreographer_postFrameworkCallback[Delayed] are deprecated since 29.
So far, calling the deprecated APIs didn't cause any warning because
compiler didn't know that they are deprecated.

As we correctly annotate the deprecation using the availability
attribute, the compiler starts to emit warning which by default is
considered as an error.

The use of the API inside the platform is considered okay because the
platform need to support the APIs for backwards compatibility.
Therefore, the errors are explicitly suppressed.

Bug: 163288375
Test: m
Change-Id: Ic9f6338e32bc6744c50ef627fbdc3f4e4afc7363