installd: rename and extract Dex2oatFileWrapper
... to UniqueFile.

Also, let the class maintain the file path as member variable. This
allows the client to access both fd and file path when they see fit.
As a result, instead of using a closure, the class now provides the path
to the cleanup function.

Additionally, reduce code duplication in constructors.

Bug: 161470356
Test: atest run_dex2oat_test

Change-Id: I1b66bc58b1885933d1601fdc2f4218c4a262d9f3