Allow re-associating NDK binder with same descriptor
An NDK binder service may get created in one module but then used in a
different module without access to the original class pointer. In this
case, the client module must try to re-associate the object with a new
class. As long as the new class's descriptor matches the object's
existing interface descriptor, we now allow the association to succeed.

Test: atest libbinder_ndk_unit_test
Change-Id: I2bb9b5189afbd124b5083daf31515fddcca2bff7