Treat an already associated Binder as remote in Rust
A binder object may have already been associated by another module (e.g.
NDK code that created the object if it is local) and therefore already
have an NDK AIBinder_Class. In this case we still want to transact with
the object, but must treat it as remote if its class does not match the
class expected for a compatible Rust service.

Bug: 167723746
Test: atest rustBinderTest
Change-Id: Id90c8153a3cc6656607c29bd8864e513c3f090d5