Fix loading ANGLE as native driver
When ANGLE namespace doesn't exist and attempt_to_load_angle is
called, ANGLE driver is loaded as built-in gl driver in sphal
namespace.

Without this, ANGLE would be loaded as a native driver without
the EGL loader knowing that it is ANGLE and cnx->useAngle and
cnx->angleBackend would never be set. This causes problems when
eglCreateWindowSurfaceTmpl() does not realize that ANGLE will
make the native_window_* calls inside ANGLE's Vulkan path and
we end up with duplicated native_window_* calls and errors like
"connect: already connected".

Bug: b/164605134
Test: boot Cuttlefish with built-in ANGLE driver
Change-Id: Ie7b80704d334c9f85187353254de615e1b78486c