Fix load_angle with built-in ANGLE driver without namespace
When ANGLE namespace doesn't exist and attempt_to_load_angle is
called, ANGLE driver is loaded as built-in gl driver in sphal
namespace.

Without this, ANGLE would be loaded as a native driver without
the EGL loader knowing that it is ANGLE and cnx->useAngle and
cnx->angleBackend would never be set. This causes problems when
eglCreateWindowSurfaceTmpl() does not realize that ANGLE will
make the native_window_* calls inside ANGLE's Vulkan path and
we end up with duplicated native_window_* calls and errors like
"connect: already connected".

Bug: b/164605134
Test: boot Cuttlefish w/ ANGLE GL + SwiftShader VK
Change-Id: I2cb9781eaff1fc94d20f0aba3ce70cfb895b4905