dumpstate: more time for non-zipping lshal debug
In a normal zipped bugreport, each individual HAL is given 2s to dump
debug information. However, there may be dozens of HALs on a device. In
order to dump all of this information when not zipping, increasing the
timeout of debugging all of the HALs to 60s.

Bug: 169025960
Test: adb bugreport > bug.txt

Merged-In: If9e8169ff15842ea4ae4a9ff3d8a641216dc93bd
Change-Id: If9e8169ff15842ea4ae4a9ff3d8a641216dc93bd