binder_parcel_fuzzer: use FuzzedDataProvider
This makes the code a little more readable and paves the way for having
more complicated Parcel-creation.

As a side effect, this allows the number of instructions to be run on a
parcel to be larger than the size of the parcel itself, so we should
increase meaningful coverage for small parcel sizes.

Bug: N/A
Test: binder_parcel_fuzzer for several minutes
Change-Id: I30a8f6c40e5255b3d01b2bacc52990cfb50150f0