InputDispatcher: fixed a minus data racing problem
Currently, there is synchronization between write and read of members of
InputApplicationInfo object, so it's possible that while InputDispatcher
thread is querying the name of the focused app, another thread changes
invokes the SurfaceControl.setInputInfo and modifies the name of the same
InputApplicationInfo object.

If we are lucky, InputDispatcher thread only gets the old name, or the
data could be corrupted and causes errors out of no reason sooner or
later.

Since InputDispatcher uses InputApplicatonHandle only to get the app
name and input anr timeout, so I think make a copy of the orignal object
wouldn't be too bad and saves use from writing mutex code.

Bug: 171727069
Test: atest -a inputflinger_tests
Test: atest ActivityManagerMultiDisplayTests
Test: atest CtsWindowManagerDeviceTestCases
Test: atest com.android.server.wm.DisplayContentTests
Test: adb logcat and inspect the InputDispatcher logs

Change-Id: I333dcf4d47f5361361930da63dfdc75d44951cb1