libbinder: -Wpadded in Parcel
This structure is used on the stack, and as such, many prebuilts have
historically assumed it is a particular size, and so it has been unsafe
to add things here. However, for b/167966510, in order to support
sockets and versioning in binder, we need approximately
sizeof(sp<IBinder>) + 1 bytes of free space in Parcel (32-bit is harder
here). This may _just_ be feasible by combining bools.

For now, make sure we understand the layout and still have even this
level of flexibility here.

Bug: 167966510
Test: build (compile time checks in place)
Change-Id: Ib7f6caaec43db92e9dcda15a695f5ba82f07e6f6