libbinder: lift parcel assoc binder ndk -> cpp
FIXME: this doesn't contain enough versioning information to support a
stable wire format, so we shouldn't consider it until it does.

FIXME: this adds a strong pointer Parcel -> transacting binder, but this
could be problematic. Alternatives include:
- storing derivative information in Parcel, instead of the binder
- storing pointer to binder instead (may leak less, but UB more, if
  there are bad clients)

FIXME: C++/Java backends need to set this information.

FIXME: tests

This CL is adding a way into Parcel to have versioning information when
reading and writing parcels, so that we can always make sure to write a
version of the wire protocol that the other side can understand.

Bug: 167966510
Test: wip
Change-Id: I313af4a01294b426030b42fbde25ad3439a40cb6