dumpstate: save background traces when a BR is captured.
[NOT FOR REVIEW yet, some details TBD]

Adds a call to perfetto --save-for-bugreport to bugreport.
If a background trace labeled with bugreport_score > 0 is
being recorded when dumpstate is invoked, the tracing daemon
will stop the trace and serialize it into a fixed location
(/data/misc/perfetto-traces/bugreport.pftrace).
If no background tracing is happening, the command is a no-op.

Bug: 170334305
Change-Id: I4278459697a5ed20b20ee99032884874a82a66be