libbinder_ndk: absorb parcelable header format
Java writeParcelable hides the 0/1 32-byte integer value from clients as
an implementation detail of the platform. It is believed that it is not
exposed anywhere else. In order to match parity with this (and perhaps,
in 5-10 years, be able to change/add the information here), absorbing
this beneath the libbinder_ndk API.

Fixes: 174091815
Test: atest CtsNdkBinderTestCases # notice, it now exercises these APIs
Change-Id: I9f3f26658c7aaac39bb51e6cdc9019eeac13e887