Includes binder bindings as a library
The generated bindings were previously included as source. This
prevents other static analyzers such as rust-analyzer to resolve the
generated symbols. Include the bindings as a rust library.

The StatusCode had to be redefined as it is not possible to define
traits on external enum (nor define traits on an enum type alias).

Bug: 175004835
Test: build
Change-Id: I0bbe6f3c62b65bced3956e566f893f3580973846