ProcessState: isThreadExpected
This adds an API to tell when there is a known thread/poller in-process
which is handling calls, and it can be used by APIs which require
threads to be handled on callbacks.

The original motivation of this was to add a log whenever a binder is
sent out. However, this causes too much logspam because it is very
common to add a service to the servicemanager and only then join the
threadpool.

Instead, a follow-up CL will abort on this variable, when it is used
incorrectly.

Bug: 175887411
Test: w/ boot + test

Change-Id: I52dcb3d36d807f0452e2cddddfa67dabed4bc002