ProcessState: isThreadExpected
Calls to binder processes which don't have threadpools started will
hang. In order to prevent this, this API is added, and a log is added
when a binder is sent out, and there is no thread available.

This adds ~30 logs to boot, and most of these places are expected to be
starting threadpools after they send out a binder (e.g. register a
service and then start a threadpool). However, this small amount of spam
is worth giving a positive hint in the case of a deadlock (especially as
HIDL developers are using AIDL).

HIDL fixes this problem by always starting a threadpool when a binder
leaves (at the HIDL level - not at the libhwbinder level). However, this
is unnacceptable for libbinder, since it would be a very backwards
incompatible change:
- the AIDL threadpool default size is up to 15 (lazy started)
- people who start the threadpool after sending out a binder in most
  cases probably need fewer threads started (or maybe even they will
  just call 'join' and have no full threadpool).

Future considerations: APIs where we expect a callback (e.g. AIDL's
waitForService) can add this as a hard error. New APIs which send out
binders can also start aborting if this condition isn't met, since
explicit threadpool creation at the start of a process is best practice.

Bug: 175887411
Test: w/ boot, manually be removing threadpool start from a test, to
    check for log, and adding case in binderLibTest

Change-Id: I52dcb3d36d807f0452e2cddddfa67dabed4bc002