Make DumpstateListener onError method oneway
Introduces a new system property for tracking whether or not
dumpstate has failed. This property will be used by the
system server to check the status of dumpstate. This will
mitigate potential race conditions between the dumpstate
binder service exiting, and the dumpstate listener receiving
an error callback.

This change allows the onError DumpstateListener call to be made
asynchronous, meaning that a blocking binder call will no longer
be made out of system_server.

Bug: 147703592
Test: Manual. Take bugreport, ensure that property is set and unset
      correctly.
Change-Id: Ibb74bf068c4b186f07e19bb076b4ddea69eda768