Keep step counter active when any client register it
[Issue]
On PRC device, the step count result is not accurate in Wechat, Alipay
and other apps.

[Root cause]
Android R implemented a new mechanism in SensorService, the sensor
hardware is disabled when all its clients go into idle mode.
This behavior is totally different from Android Q.
On Android Q, when all sensor clients go into idle mode, only the
sensor events are suspended sending to the clients, the sensor
hardware itself is not disabled.

However, according to Android API doc for the step counter:
https://developer.android.com/reference/android/hardware/Sensor#TYPE_STEP_COUNTER
...If you want to continuously track the number of steps over a long
period of time, do NOT unregister for this sensor, so that it keeps
counting steps in the background even when the AP is in suspend mode and
report the aggregate count when the AP is awake. Application needs to
stay registered for this sensor because step counter does not count
steps if it is not activated....

This mechanism on Android R does not meet the use case of step counter.

[Solution]
Per the special use case of step counter, we should keep the step
counter hardware active when any app register it, regardless of the
app state being active or idle.


Signed-off-by: Shuxian Zhang <zhangsx1@motorola.com>
Change-Id: I8880ca5d9e6e16c37951302273eda1689065b9c6