Avoid integer underflow involving maxFreqCount.
If maxFreqCount is zero, as is the case on emulators that do not
support frequency scaling, we would underflow when subtracting 1 and
we may end up spending a significant amount of time in these loops,
likely leading to an ANR in system_server since at least the loop in
clearUidTimes is executed while holding the BatteryStatsImpl lock. Fix
the problem by not entering the loops if maxFreqCount is zero.

Change-Id: I2ec1e8de0fe2a40ed100c8f14e6ca3f6d6285b82