lshal: emphasize HIDL in docs/output
Some IMs floating about this recently. For now, lshal is still HIDL
specific and binder-facing tools work with AIDL HALs.

It remains to be seen if lshal will be re-adapted for AIDL in the
future, or if some of the ideas here will be transfered to dumpsys and
this binary ultimately deprecated.

Bug: N/A
Test: N/A
Change-Id: I39be9533b44c085a0688eb39171a0a40f88067b2