Fix sensor actvie after SensorEventConnection removed
When app idle and register a sensor on existing connecion,the method
sensor.batch() retrun NO_ERROR but sensor.activite() return INVALID_OPERATION.
The sensor will remove from the connecion but unable to remove from the
batchParams,this make the sensor active after the app killed.
As the client is disabled, keep the sensor in the connection for resume.

Bug: 180825144

Test:debug app in 180825144

Signed-off-by: xiamengsen <xiamengsen@xiaomi.com>
Change-Id: Icc22862d1a083abba08fb2cbcabd043444b7090f