dumpstate: Use UTC instead of local time
This commit make dumpstate and dumpsys use UTC and ISO-8601
for timestamps with making the timezone explicit.

Timestamps in the output of dumpstate and dumpsys used the local
time without clarifying the timezone. The output may be used in
places where UTC is mainly used, especially Chrome OS, and the
timestamps are very confusing in such places.

Bug: 180562941
Test: atest dumpsys_test
Test: atest dumpstate_test
Test: atest dumpstate_smoke_test
Test: run dumpstate and check that all timestamps use UTC
Change-Id: I3d0e2cffe615443a21f736f8dd2463a20e9af344