Check if the buffer is actually being scaled instead of only checking scaling mode
BufferStateLayers have a default scale to window scaling mode which
means the layers would be set use texture filtering regardless of
the buffer size. This was breaking some pixel by pixel screenshot
comparison tests. Instead check if the buffer size, after applying
any buffer transforms, matches the layer size.

This is a cherry pick of ag/12947725, preparing for a subsequent change
to apply the same logic to BufferLayer.

Bug: b/182584062
Test: android.view.inputmethod.cts.FocusHandlingTest#testNonFocusablePopupWindowDoesNotAffectImeVisibility
Change-Id: I8a64ad96c82b65abf5d1312447ff76915c61f5d5