libmemtrackproxy
Helper library that implements memtrack proxy service. The service
forwards requests to the appropriate AIDL or HIDL memtrack HAL if
the following access control requirements are
met:
  - Only AID_SYSTEM or AID_ROOT can request getMemory() for other
    processes, otherwise the requested PID must match the calling
    PID.
  - Only AID_SYSTEM or AID_ROOT can request getGpuDeviceInfo()

If the above conditions are not met, the request is not forwarded
to the memtrack HAL and a status with exception code EX_SECURITY
is returned.

Bug: 177664629
Test: memtrackproxy_test
Change-Id: Icf3e0fa0500b2961b0945e800899881c9a3c86cf