[DO NOT SUBMIT] testing DropRootUser in dumpsys
DropRootUser returns true, but does not cause the following test case to
fail. If the following test case should fail without root access.

I tried with DropRootUser being called in different places with no
change in behavior.
I tried with DropRootUser being called multipls times, and it fails on
the subsequent calls. So it must be doing _something_.

Test: atest dumpsys_test

Change-Id: I69bad885cec32a8e2bab92f61cccbdead60a612f