Finish bugreportz when OK: is received.
Currently bugreportz finishes only when pipe is closed,
and read() returns 0 bytes.
But on some devices read() hangs forever and doesn't exit -
even if the other side closed the pipe and the process is exited.

As dumpstate is guaranteed to return OK: on successful
zip file creation, bugreportz should leverage this behavior.

Test: m -j
Test: flash and run "adb bugreport"
Bug: 184211736
Change-Id: I98bfe1d348acc4d64da383db38f1ac567011040b