Split libbinder into unrelated sublibraries.
This is in preparation to remove these sublibraries, which add a fixed
cost to using libbinder which is unnecessary for most libbinder
users.

With merge breaks, since each of these can be removed simultaneously,
and if they are like battery stats, may require multiple days and
multiple CLs in multiple gerrits to land, this way there is no ordering
dependency on this work.

This also removed an unused set of interfaces from the vendor variant of
libbinder.

Bug: 183654927
Test: build & boot
Change-Id: I21b4320e4560346804e9f07d9bfadae16a25aa2e