Call apexd --unmount-all at the end of otapreopt_chroot run
We don't really care for the umount part, but we do care about freeing
up loop and dm block devices so that next run of otapreopt_chroot can
re-use them.

Note that we must run `apexd --unmount-all` inside otapreopt_chroot
itself. This is because otapreopt_chroot runs in /postintall namespace.

Test: install an OTA, check number of loop devices
Bug: 184716804
Change-Id: I4f01cb9082f90fd40fba4a1864e8d6b548f5a274