Fix the secure flag check for blackout layer issue
When there is a layer not to be composed on insecure virtual display, surfaceflinger will send a 
black color layer to render engine to compose instead of the original buffer.

If render engine supports protected context, but layer does not set the eLayerSecure flag,the layer
will not be set to black color layer and will be sent to render engine to compose on the insecure 
display based on current logic, even if the layer buffer usage is protected.

Solution: add buffer usage check for insecure display case.
Test result: The secure layer buffer will not be sent to render engine for the insecure virtual 
display to compose, and a black color layer will be sent instead.

Change-Id: Iad53411ea642fee3e5b09222760a223b4ada61ad